refactor: footer 수정
